+++ src/App.tsx
@@ -1,21 +1,15 @@
-// import { useState } from "react";
-// import reactLogo from "./assets/react.svg";
-// import viteLogo from "/vite.svg";
-// import "./App.css";
 import { Toaster } from "react-hot-toast";
 import { AuthProvider } from "./context/auth-context";
 import Routers from "./routes/router";
 
 function App() {
-  // const [count, setCount] = useState(0);
-
   return (
-    <>
-      <AuthProvider>
+    <AuthProvider>
+      <>
         <Toaster />
         <Routers />
-      </AuthProvider>
-    </>
+      </>
+    </AuthProvider>
   );
 }
